Feature/review

+++ clients/src/views/userview/TestProductDetail.vue
@@ -450,7 +450,6 @@ export default {
             }
         },
         async showReviewList(pageno) {
-                console.log(pageno);
                 this.$showLoading();
                 const result = await axios.get(`/api/product/productdetails/review/${this.product_no}/${pageno}`)
                     .catch((err) =>
@@ -471,6 +470,11 @@ export default {
                 this.$hideLoading();
             },
             async addReviewLikeCnt(rno) {
+                if(this.$store.state.userNo <= 0) {
+                    this.$showWarningAlert('로그인을 해주세요.');
+                    return;
+                }
+
                 this.$showLoading();
                 const result = await axios.put(`/api/product/productdetails/review/${rno}`)
                     .catch((err) => console.log(err));
@@ -479,6 +483,11 @@ export default {
                 this.showReviewList(this.page.curPage);
             },
             async cancleReviewLikeCnt(rno) {
+                if(this.$store.state.userNo <= 0) {
+                    this.$showWarningAlert('로그인을 해주세요.');
+                    return;
+                }
+
                 this.$showLoading();
                 const result = await axios.delete(`/api/product/productdetails/review/${rno}`)
                     .catch((err) => console.log(err));
